ECOM-034: Migrate monetary fields from Double to BigDecimal across Product, Cart, Order and Analytics modules

+++ src/main/java/com/ecommerce/project/payload/CreateProductRequest.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -23,10 +25,10 @@ public class CreateProductRequest {
 
     @NotNull
     @Positive
-    private Double price;
+    private BigDecimal price;
 
     @NotNull
     @DecimalMin("0.0")
     @DecimalMax("100.0")
-    private Double discount;
+    private BigDecimal discount;
 }
