+++ src/main/java/com/ecommerce/project/payload/OrderItemResponseDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -21,7 +23,7 @@ public class OrderItemResponseDTO {
 
     private Integer quantityOrdered;
 
-    private Double orderedProductPrice;
+    private BigDecimal orderedProductPrice;
 
-    private Double discount;
+    private BigDecimal discount;
 }
