+++ src/main/java/com/ecommerce/project/payload/ProductDTO.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -26,12 +28,12 @@ public class ProductDTO {
 
     @NotNull
     @Positive
-    private Double price;
+    private BigDecimal price;
 
     @NotNull
     @DecimalMin("0.0")
     @DecimalMax("100.0")
-    private Double discount;
+    private BigDecimal discount;
 
-    private Double specialPrice;
+    private BigDecimal specialPrice;
 }
